refactor(redteam): split oversized seed.ts into 6 per-concern modules (#1047)
apps/web/e2e/redteam/helpers/seed.ts was 670 lines — 466 over the 200-line
utility cap (code-style.md §1) and growing monotonically as each new sub-vector
appended constants. Split it by concern into 6 sibling modules and delete the
original outright (no re-export barrel, per code-style.md §4):

- seed-core.ts     (81)  — OTHER_ORG_SLUG, getEgmontOrgId, upsertUser (spine)
- seed-users.ts    (198) — creds consts + the 6 seed*User orchestration fns
- seed-quiz.ts     (170) — pickSubjectWithQuestions, ensureExamConfig
- seed-responses.ts(144) — VictimResponseFixture, seedVictimResponses
- seed-consent.ts  (54)  — USER_CONSENTS_* + seedConsentRecords
- seed-markers.ts  (25)  — all E2E_* fixture marker constants

All code moved verbatim (byte-identical bodies; impl-critic confirmed). The 48
spec importers + get-study-questions-eo-setup.ts repointed to the specific new
modules; seed.test.ts renamed to seed-quiz.test.ts (co-located with its subject).

Test-infra only — no production runtime touched. tsc 0, biome clean, the moved
unit test passes 7/7.

Closes #1047
