+++ apps/web/e2e/redteam/quiz-draft-injection.spec.ts
@@ -19,15 +19,15 @@
 import { expect, test } from '@playwright/test'
 import { getAdminClient } from '../helpers/supabase'
 import { createAuthenticatedClient } from './helpers/redteam-client'
+import { pickSubjectWithQuestions } from './helpers/seed-quiz'
 import {
   ATTACKER_EMAIL,
   ATTACKER_PASSWORD,
   createCrossOrgUser,
-  pickSubjectWithQuestions,
   seedRedTeamUsers,
   VICTIM_EMAIL,
   VICTIM_PASSWORD,
-} from './helpers/seed'
+} from './helpers/seed-users'
 
 test.describe('Red Team: Quiz Draft Question Injection', () => {
   let attackerClient: Awaited<ReturnType<typeof createAuthenticatedClient>>
