+++ apps/web/e2e/redteam/quiz-session-score-forgery.spec.ts
@@ -22,12 +22,8 @@
 import { expect, test } from '@playwright/test'
 import { getAdminClient } from '../helpers/supabase'
 import { createAuthenticatedClient } from './helpers/redteam-client'
-import {
-  ATTACKER_EMAIL,
-  ATTACKER_PASSWORD,
-  pickSubjectWithQuestions,
-  seedRedTeamUsers,
-} from './helpers/seed'
+import { pickSubjectWithQuestions } from './helpers/seed-quiz'
+import { ATTACKER_EMAIL, ATTACKER_PASSWORD, seedRedTeamUsers } from './helpers/seed-users'
 
 test.describe('Red Team: quiz_sessions Score Forgery (direct UPDATE)', () => {
   let attackerClient: Awaited<ReturnType<typeof createAuthenticatedClient>>
