+++ apps/web/e2e/redteam/rpc-admin-dashboard-students.spec.ts
@@ -24,7 +24,7 @@ import {
   seedRedTeamAdmin,
   seedRedTeamInstructor,
   seedRedTeamUsers,
-} from './helpers/seed'
+} from './helpers/seed-users'
 
 const SUPABASE_URL = process.env.NEXT_PUBLIC_SUPABASE_URL ?? 'http://localhost:54321'
 const ANON_KEY = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY ?? ''
