+++ apps/web/e2e/redteam/rpc-internal-exam-admin-soft-deleted.spec.ts
@@ -39,7 +39,8 @@
 import { expect, test } from '@playwright/test'
 import { getAdminClient } from '../helpers/supabase'
 import { createAuthenticatedClient } from './helpers/redteam-client'
-import { ensureExamConfig, pickSubjectWithQuestions, seedRedTeamUsers } from './helpers/seed'
+import { ensureExamConfig, pickSubjectWithQuestions } from './helpers/seed-quiz'
+import { seedRedTeamUsers } from './helpers/seed-users'
 
 // Dedicated throwaway admin — NOT the shared redteam-admin@lmsplus.local.
 // Soft-deleting the shared admin would cascade not_admin failures across all 12
