+++ apps/web/e2e/redteam/rpc-report-answer-keys.spec.ts
@@ -34,16 +34,18 @@ import { createClient } from '@supabase/supabase-js'
 import { getAdminClient } from '../helpers/supabase'
 import { createAuthenticatedClient } from './helpers/redteam-client'
 import {
-  ATTACKER_EMAIL,
-  ATTACKER_PASSWORD,
   E2E_REDTEAM_EN_MARKER,
   E2E_REDTEAM_EN_SOFTDEL_STUDENT_EMAIL,
   E2E_REDTEAM_EN_SOFTDEL_STUDENT_PASSWORD,
-  pickSubjectWithQuestions,
+} from './helpers/seed-markers'
+import { pickSubjectWithQuestions } from './helpers/seed-quiz'
+import {
+  ATTACKER_EMAIL,
+  ATTACKER_PASSWORD,
   seedRedTeamUsers,
   VICTIM_EMAIL,
   VICTIM_PASSWORD,
-} from './helpers/seed'
+} from './helpers/seed-users'
 
 const SUPABASE_URL = process.env.NEXT_PUBLIC_SUPABASE_URL ?? 'http://localhost:54321'
 const ANON_KEY = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY
