+++ apps/web/e2e/redteam/users-role-forge.spec.ts
@@ -39,7 +39,7 @@
 import { expect, test } from '@playwright/test'
 import { getAdminClient } from '../helpers/supabase'
 import { createAuthenticatedClient } from './helpers/redteam-client'
-import { ATTACKER_EMAIL, ATTACKER_PASSWORD, seedRedTeamUsers } from './helpers/seed'
+import { ATTACKER_EMAIL, ATTACKER_PASSWORD, seedRedTeamUsers } from './helpers/seed-users'
 
 test.describe('Red Team: users Role Forge (privilege escalation via direct UPDATE)', () => {
   let studentClient: Awaited<ReturnType<typeof createAuthenticatedClient>>
